requirements.txt
Django==1.6.2
Pillow==2.3.1
amqp==1.4.4
anyjson==0.3.3
billiard==3.3.0.16
boto==2.27.0
celery==3.1.10
django-appconf==0.6
django-compressor==1.3
django-crispy-forms==1.4.0
django-endless-pagination==2.0
django-registration==1.0
django-storages==1.1.8
django-taggit==0.11.2
django-widget-tweaks==1.3
djangorestframework==2.3.10
easypost==2.0.7
firebase-token-generator==1.3.2
gunicorn==18.0
ipython==1.2.1
kombu==3.0.14
oauthlib==0.6.1
pystache==0.5.3
-e git://github.com/mikexstudios/python-firebase.git@fc0597aafb1cb83c76e039370b1e0e180322f0e0#egg=python_firebase-master
python-openid==2.2.5
python-social-auth==0.1.22
pytz==2014.2
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.3.3
six==1.6.1
sqlparse==0.1.11
static==1.0.2
stripe==1.12.2
wsgiref==0.1.2
#psycopg2==2.5.2
dj-database-url==0.3.0
newrelic
